refactor: move changes to vaadin-date-picker package
