--- packages/date-picker/README.md
@@ -1,5 +1,0 @@
-# @vaadin/date-picker
-
-> ⚠️ Work in progress, please do not use this component yet.
-
-The new version of `vaadin-date-picker` component using slotted `<input>`.
