--- packages/date-picker/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@vaadin/date-picker",
-  "version": "22.0.0-alpha5",
-  "description": "vaadin-date-picker",
-  "main": "vaadin-date-picker.js",
-  "module": "vaadin-date-picker.js",
-  "repository": "vaadin/web-components",
-  "keywords": [
-    "Vaadin",
-    "date-picker",
-    "web-components",
-    "web-component"
-  ],
-  "author": "Vaadin Ltd",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/vaadin/web-components/issues"
-  },
-  "homepage": "https://vaadin.com/components",
-  "files": [
-    "vaadin-*.d.ts",
-    "vaadin-*.js",
-    "src",
-    "theme"
-  ],
-  "dependencies": {
-    "@polymer/polymer": "^3.0.0",
-    "@vaadin/field-base": "^22.0.0-alpha5",
-    "@vaadin/input-container": "^22.0.0-alpha5",
-    "@vaadin/text-field": "^22.0.0-alpha5",
-    "@vaadin/vaadin-date-picker": "^22.0.0-alpha5",
-    "@vaadin/vaadin-element-mixin": "^22.0.0-alpha5",
-    "@vaadin/vaadin-lumo-styles": "^22.0.0-alpha5",
-    "@vaadin/vaadin-material-styles": "^22.0.0-alpha5",
-    "@vaadin/vaadin-themable-mixin": "^22.0.0-alpha5"
-  },
-  "devDependencies": {
-    "@esm-bundle/chai": "^4.3.4",
-    "@vaadin/testing-helpers": "^0.2.1",
-    "sinon": "^9.2.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
