--- packages/date-picker/src/vaadin-date-picker.js
@@ -1,150 +1,0 @@
-/**
- * @license
- * Copyright (c) 2021 Vaadin Ltd.
- * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
- */
-import { PolymerElement, html } from '@polymer/polymer';
-import { GestureEventListeners } from '@polymer/polymer/lib/mixins/gesture-event-listeners.js';
-import '@polymer/iron-media-query/iron-media-query.js';
-import { ElementMixin } from '@vaadin/vaadin-element-mixin/vaadin-element-mixin.js';
-import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { ClearButtonMixin } from '@vaadin/field-base/src/clear-button-mixin.js';
-import { DelegateFocusMixin } from '@vaadin/field-base/src/delegate-focus-mixin.js';
-import { FieldAriaMixin } from '@vaadin/field-base/src/field-aria-mixin.js';
-import { InputConstraintsMixin } from '@vaadin/field-base/src/input-constraints-mixin.js';
-import { InputSlotMixin } from '@vaadin/field-base/src/input-slot-mixin.js';
-import '@vaadin/vaadin-date-picker/src/vaadin-date-picker-overlay.js';
-import '@vaadin/vaadin-date-picker/src/vaadin-date-picker-overlay-content.js';
-import '@vaadin/input-container/src/vaadin-input-container.js';
-import '@vaadin/text-field/src/vaadin-input-field-shared-styles.js';
-import { DatePickerMixin } from './vaadin-date-picker-mixin.js';
-
-export class DatePicker extends DatePickerMixin(
-  FieldAriaMixin(
-    ClearButtonMixin(
-      InputConstraintsMixin(
-        InputSlotMixin(DelegateFocusMixin(GestureEventListeners(ThemableMixin(ElementMixin(PolymerElement)))))
-      )
-    )
-  )
-) {
-  static get is() {
-    return 'vaadin-date-picker';
-  }
-
-  static get template() {
-    return html`
-      <style include="vaadin-input-field-shared-styles">
-        :host([dir='rtl']) [part='input-field'] {
-          direction: ltr;
-        }
-
-        :host([dir='rtl']) [part='input-field'] ::slotted(input)::placeholder {
-          direction: rtl;
-          text-align: left;
-        }
-
-        :host([opened]) {
-          pointer-events: auto;
-        }
-      </style>
-
-      <div part="container">
-        <div part="label" on-click="focus">
-          <slot name="label"></slot>
-          <span part="indicator" aria-hidden="true"></span>
-        </div>
-
-        <vaadin-input-container
-          part="input-field"
-          readonly="[[readonly]]"
-          disabled="[[disabled]]"
-          invalid="[[invalid]]"
-          theme$="[[theme]]"
-        >
-          <slot name="prefix" slot="prefix"></slot>
-          <slot name="input"></slot>
-          <div id="clearButton" part="clear-button" slot="suffix"></div>
-          <div part="toggle-button" slot="suffix" on-tap="_toggle" role="button"></div>
-        </vaadin-input-container>
-
-        <div part="helper-text">
-          <slot name="helper"></slot>
-        </div>
-
-        <div part="error-message">
-          <slot name="error-message"></slot>
-        </div>
-      </div>
-
-      <vaadin-date-picker-overlay
-        id="overlay"
-        fullscreen$="[[_fullscreen]]"
-        theme$="[[__getOverlayTheme(theme, _overlayInitialized)]]"
-        on-vaadin-overlay-open="_onOverlayOpened"
-        on-vaadin-overlay-close="_onOverlayClosed"
-        disable-upgrade
-      >
-        <template>
-          <vaadin-date-picker-overlay-content
-            id="overlay-content"
-            i18n="[[i18n]]"
-            fullscreen$="[[_fullscreen]]"
-            label="[[label]]"
-            selected-date="{{_selectedDate}}"
-            slot="dropdown-content"
-            focused-date="{{_focusedDate}}"
-            show-week-numbers="[[showWeekNumbers]]"
-            min-date="[[_minDate]]"
-            max-date="[[_maxDate]]"
-            role="dialog"
-            on-date-tap="_close"
-            part="overlay-content"
-            theme$="[[__getOverlayTheme(theme, _overlayInitialized)]]"
-          ></vaadin-date-picker-overlay-content>
-        </template>
-      </vaadin-date-picker-overlay>
-
-      <iron-media-query query="[[_fullscreenMediaQuery]]" query-matches="{{_fullscreen}}"> </iron-media-query>
-    `;
-  }
-
-  static get properties() {
-    return {
-      /** @private */
-      _userInputValue: String
-    };
-  }
-
-  static get observers() {
-    return ['_userInputValueChanged(_userInputValue)'];
-  }
-
-  /**
-   * Used by `ClearButtonMixin` as a reference to the clear button element.
-   * @protected
-   * @return {!HTMLElement}
-   */
-  get clearElement() {
-    return this.$.clearButton;
-  }
-
-  /** @private */
-  _onVaadinOverlayClose(e) {
-    if (this._openedWithFocusRing && this.hasAttribute('focused')) {
-      this.setAttribute('focus-ring', '');
-    } else if (!this.hasAttribute('focused')) {
-      // TODO: consider preserving focus
-      this.blur();
-    }
-    if (e.detail.sourceEvent && e.detail.sourceEvent.composedPath().indexOf(this) !== -1) {
-      e.preventDefault();
-    }
-  }
-
-  /** @private */
-  _toggle(e) {
-    e.stopPropagation();
-    this[this._overlayInitialized && this.$.overlay.opened ? 'close' : 'open']();
-  }
-}
