--- packages/date-picker/vaadin-date-picker.js
@@ -1,1 +1,0 @@
-// export { TextArea } from './src/vaadin-text-area.js';
