+++ packages/vaadin-date-picker/README.md
@@ -1,65 +1,5 @@
-# &lt;vaadin-date-picker&gt;
+# @vaadin/date-picker
 
-[Live Demo ↗](https://vaadin.com/components/vaadin-date-picker/html-examples)
-|
-[API documentation ↗](https://vaadin.com/components/vaadin-date-picker/html-api)
+> ⚠️ Work in progress, please do not use this component yet.
 
-[&lt;vaadin-date-picker&gt;](https://vaadin.com/components/vaadin-date-picker) is a Web Component providing a date selection field which includes a scrollable month calendar view, part of the [Vaadin components](https://vaadin.com/components).
-
-[![npm version](https://badgen.net/npm/v/@vaadin/vaadin-date-picker)](https://www.npmjs.com/package/@vaadin/vaadin-date-picker)
-[![Published on Vaadin Directory](https://img.shields.io/badge/Vaadin%20Directory-published-00b4f0.svg)](https://vaadin.com/directory/component/vaadinvaadin-date-picker)
-[![Discord](https://img.shields.io/discord/732335336448852018?label=discord)](https://discord.gg/PHmkCKC)
-
-```html
-<vaadin-date-picker label="Label" placeholder="Placeholder"></vaadin-date-picker>
-```
-
-[<img src="https://raw.githubusercontent.com/vaadin/vaadin-date-picker/master/screenshot.png" width="439" alt="Screenshot of vaadin-date-picker">](https://vaadin.com/components/vaadin-date-picker)
-
-## Installation
-
-Install `vaadin-date-picker`:
-
-```sh
-npm i @vaadin/vaadin-date-picker --save
-```
-
-Once installed, import it in your application:
-
-```js
-import '@vaadin/vaadin-date-picker/vaadin-date-picker.js';
-```
-
-## Getting started
-
-Vaadin components use the Lumo theme by default.
-
-To use the Material theme, import the correspondent file from the `theme/material` folder.
-
-## Entry points
-
-- The components with the Lumo theme:
-
-  `theme/lumo/vaadin-date-picker.js`
-  `theme/lumo/vaadin-date-picker-light.js`
-
-- The components with the Material theme:
-
-  `theme/material/vaadin-date-picker.js`
-  `theme/material/vaadin-date-picker-light.js`
-
-- Alias for `theme/lumo/vaadin-date-picker.js`
-  `theme/lumo/vaadin-date-picker-light.js`:
-
-  `vaadin-date-picker.js`
-  `vaadin-date-picker-light.js`
-
-## Contributing
-
-Read the [contributing guide](https://vaadin.com/docs/latest/guide/contributing/overview) to learn about our development process, how to propose bugfixes and improvements, and how to test your changes to Vaadin components.
-
-## License
-
-Apache License 2.0
-
-Vaadin collects development time usage statistics to improve this product. For details and to opt-out, see https://github.com/vaadin/vaadin-usage-statistics.
+The new version of `vaadin-date-picker` component using slotted `<input>`.
