+++ packages/vaadin-date-picker/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@vaadin/vaadin-date-picker",
+  "name": "@vaadin/date-picker",
   "version": "22.0.0-alpha6",
-  "description": "Web component providing a date selection field with scrollable month calendar",
+  "description": "vaadin-date-picker",
   "main": "vaadin-date-picker.js",
   "module": "vaadin-date-picker.js",
-  "repository": "vaadin/vaadin-date-picker",
+  "repository": "vaadin/web-components",
   "keywords": [
     "Vaadin",
     "date-picker",
@@ -14,7 +14,7 @@
   "author": "Vaadin Ltd",
   "license": "Apache-2.0",
   "bugs": {
-    "url": "https://github.com/vaadin/vaadin-date-picker/issues"
+    "url": "https://github.com/vaadin/web-components/issues"
   },
   "homepage": "https://vaadin.com/components",
   "files": [
@@ -24,26 +24,20 @@
     "theme"
   ],
   "dependencies": {
-    "@polymer/iron-a11y-announcer": "^3.0.0",
-    "@polymer/iron-a11y-keys-behavior": "^3.0.0",
-    "@polymer/iron-media-query": "^3.0.0",
-    "@polymer/iron-resizable-behavior": "^3.0.0",
-    "@polymer/polymer": "^3.2.0",
-    "@vaadin/button": "^22.0.0-alpha6",
+    "@polymer/polymer": "^3.0.0",
     "@vaadin/component-base": "^22.0.0-alpha6",
-    "@vaadin/vaadin-control-state-mixin": "^22.0.0-alpha6",
+    "@vaadin/field-base": "^22.0.0-alpha6",
+    "@vaadin/input-container": "^22.0.0-alpha6",
+    "@vaadin/text-field": "^22.0.0-alpha6",
+    "@vaadin/vaadin-date-picker": "^22.0.0-alpha6",
     "@vaadin/vaadin-lumo-styles": "^22.0.0-alpha6",
     "@vaadin/vaadin-material-styles": "^22.0.0-alpha6",
-    "@vaadin/vaadin-overlay": "^22.0.0-alpha6",
-    "@vaadin/vaadin-text-field": "^22.0.0-alpha6",
     "@vaadin/vaadin-themable-mixin": "^22.0.0-alpha6"
   },
   "devDependencies": {
     "@esm-bundle/chai": "^4.3.4",
-    "@vaadin/testing-helpers": "^0.3.0",
-    "@vaadin/vaadin-dialog": "^22.0.0-alpha6",
-    "@vaadin/vaadin-template-renderer": "^22.0.0-alpha6",
-    "sinon": "^9.2.0"
+    "@vaadin/testing-helpers": "^0.2.1",
+    "sinon": "^9.2.1"
   },
   "publishConfig": {
     "access": "public"
