+++ packages/vaadin-date-picker/test/dialog.test.js
@@ -2,19 +2,18 @@ import { expect } from '@esm-bundle/chai';
 import { fixtureSync, nextFrame } from '@vaadin/testing-helpers';
 import '@vaadin/vaadin-dialog/vaadin-dialog.js';
 import '@vaadin/vaadin-template-renderer';
-import './not-animated-styles.js';
-import '../vaadin-date-picker.js';
+import './vaadin-date-picker.js';
 
 describe('modeless dialog', () => {
   let dialog, datepicker;
 
   beforeEach(async () => {
     dialog = fixtureSync(`
-    <vaadin-dialog modeless>
-      <template>
-        <vaadin-date-picker></vaadin-date-picker>
-      </template>
-    </vaadin-dialog>
+      <vaadin-dialog modeless>
+        <template>
+          <vaadin-date-picker></vaadin-date-picker>
+        </template>
+      </vaadin-dialog>
     `);
     dialog.opened = true;
     await nextFrame();
