+++ packages/vaadin-date-picker/vaadin-date-picker.d.ts
@@ -1,2 +1,1 @@
-export * from './src/vaadin-date-picker.js';
-export * from './src/interfaces';
+// export { TextArea } from './src/vaadin-text-area.js';
