+++ packages/vaadin-date-picker/vaadin-date-picker.js
@@ -1,2 +1,1 @@
-import './theme/lumo/vaadin-date-picker.js';
-export * from './src/vaadin-date-picker.js';
+// export { TextArea } from './src/vaadin-text-area.js';
